feat: add first performance tests
This is initial commit with initial test for only one object sending many messages. The difficulty of testing many nodes is, that it requires classes with different object id which needs:
either many different interface classes or editing the generated code. ALSO Tests are not build with CI, but at some point they should be added. The downside is that soulution need to be regenarated  (although otherwise they might got old) To use it under the visual studio performance profiler I needed to change generated cmakes to fetch poco lib (which is not delivered)