build: centralize deps mngt in cmakelists
The root CMakeLists has a new option APIGEAR_FETCH_DEPS. If enabled,
cmake checks for the dependencies, builds and installs them if needed.
Otherwise, all dependencies need to be pre-installed.
The apigear module has an additional option APIGEAR_FETCH_OLINKCORE,
which is disabled by default and should only be used for conan
builds.