docs: update info on starting olink server
Adds information that server should be started when all services are
added to registry.
If client wants to make a connection before the server is up, the olink
client will reconnect until it will succeed. On server start, connection
gets established, but if a services is not already to service, the Olink
will fail to link service side with a client side. To avoid such
sitiation the services should be added to registry befor server starts
listening.